Refactor TabbedForms
