+++ plugin-hrm-form/src/components/tabbedForms/BranchingFormIssueCategory.jsx
@@ -2,8 +2,8 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { Checkbox } from '@material-ui/core';
 
-import { CategoryCheckboxField, StyledCheckboxLabel, StyledLabel } from '../Styles/HrmStyles';
-import { formType } from '../types';
+import { CategoryCheckboxField, StyledCheckboxLabel, StyledLabel } from '../../Styles/HrmStyles';
+import { formType } from '../../types';
 
 const BranchingFormIssueCategory = props => {
   const cat = props.category;
