+++ plugin-hrm-form/src/components/tabbedForms/index.js
@@ -1,0 +1,3 @@
+import TabbedForms from './TabbedForms';
+
+export default TabbedForms;
